+++ src/test/java/rx/RxPublisherTest.java
@@ -7,7 +7,7 @@ import org.testng.annotations.Test;
 import rx.test.IterableDecrementer;
 
 @Test // needed for Gradle to find this as a test
-public class RxPublisherTest extends PublisherVerification<Integer> {
+public class RxPublisherTest extends PublisherVerification<Long> {
 
     public static final long DEFAULT_TIMEOUT_MILLIS = 300L;
     public static final long PUBLISHER_REFERENCE_CLEANUP_TIMEOUT_MILLIS = 1000L;
@@ -22,13 +22,13 @@ public class RxPublisherTest extends PublisherVerification<Integer> {
     }
 
     @Override
-    public Publisher<Integer> createPublisher(long elements) {
+    public Publisher<Long> createPublisher(long elements) {
         return RxReactiveStreams.toPublisher(Observable.from(new IterableDecrementer(elements)));
     }
 
     @Override
-    public Publisher<Integer> createErrorStatePublisher() {
-        return RxReactiveStreams.toPublisher(Observable.<Integer>error(new Exception("!")));
+    public Publisher<Long> createErrorStatePublisher() {
+        return RxReactiveStreams.toPublisher(Observable.<Long>error(new Exception("!")));
     }
 
 }
