+++ src/test/java/rx/RxSubscriberBlackboxTest.java
@@ -10,7 +10,7 @@ import rx.observers.Subscribers;
 import rx.test.IterableDecrementer;
 
 @Test
-public class RxSubscriberBlackboxTest extends SubscriberBlackboxVerification<Integer> {
+public class RxSubscriberBlackboxTest extends SubscriberBlackboxVerification<Long> {
 
     public static final long DEFAULT_TIMEOUT_MILLIS = 300L;
 
@@ -19,12 +19,12 @@ public class RxSubscriberBlackboxTest extends SubscriberBlackboxVerification<Int
     }
 
     @Override
-    public Subscriber<Integer> createSubscriber() {
-        return new PublisherSubscriber<Integer>(Subscribers.empty());
+    public Subscriber<Long> createSubscriber() {
+        return new PublisherSubscriber<Long>(Subscribers.empty());
     }
 
     @Override
-    public Publisher<Integer> createHelperPublisher(long elements) {
+    public Publisher<Long> createHelperPublisher(long elements) {
         return RxReactiveStreams.toPublisher(Observable.from(new IterableDecrementer(elements)));
     }
